server/requirements.txt
flask
flask_cors
waitress
python-dotenv
opencv-python
pillow
tensorflow